flask_app/requirements.txt
flask
np
numpy
Pillow
gunicorn
